+++ src/data/repositories/speach_to_text_repository_impl.py
@@ -7,8 +7,8 @@ from fastapi import Depends
 from core.config.app_config import AppConfig
 from core.logger.logger import Logger
 from core.timer.timer import TimerFactory
+from data.factories.speach_to_text_worker_factory import SpeachToTextWorkerFactory
 from data.repositories.directory_repository_impl import DirectoryRepositoryImpl
-from data.workers.whisper_worker import WhisperConfig, WhisperWorker
 from domain.repositories.directory_repository import DirectoryRepository
 from domain.repositories.speach_to_text_repository import SpeachToTextRepository
 
@@ -23,13 +23,13 @@ class SpeachToTextRepositoryImpl(SpeachToTextRepository):  # type: ignore
         directory_repository: Annotated[DirectoryRepository, Depends(DirectoryRepositoryImpl)],
         timer_factory: Annotated[TimerFactory, Depends()],
         logger: Annotated[Logger, Depends()],
-        worker: Annotated[WhisperWorker, Depends()],
+        worker_factory: Annotated[SpeachToTextWorkerFactory, Depends()],
     ) -> "SpeachToTextRepositoryImpl":
         if cls._instance is None:
             with cls._lock:
                 if cls._instance is None:
                     cls._instance = super(SpeachToTextRepositoryImpl, cls).__new__(cls)
-                    cls._instance._initialize(config, directory_repository, timer_factory, logger, worker)
+                    cls._instance._initialize(config, directory_repository, timer_factory, logger, worker_factory)
         return cls._instance
 
     def _initialize(
@@ -38,13 +38,13 @@ class SpeachToTextRepositoryImpl(SpeachToTextRepository):  # type: ignore
         directory_repository: DirectoryRepository,
         timer_factory: TimerFactory,
         logger: Logger,
-        worker: WhisperWorker,
+        worker_factory: SpeachToTextWorkerFactory,
     ) -> None:
         directory_repository.create_directory(config.speach_to_text_model_download_path)
         self.config = config
         self.timer = timer_factory.create()
         self.logger = logger
-        self.worker = worker
+        self.worker = worker_factory.create()
         self.last_access_time = 0.0
 
     def transcribe(
@@ -54,13 +54,7 @@ class SpeachToTextRepositoryImpl(SpeachToTextRepository):  # type: ignore
     ) -> dict[str, str]:
         with self._lock:
             if not self.worker.is_alive():
-                self.worker.start(
-                    WhisperConfig(
-                        device=self.config.device,
-                        model_type=self.config.speach_to_text_model_type,
-                        model_download_path=self.config.speach_to_text_model_download_path,
-                    )
-                )
+                self.worker.start()
 
         result: dict[str, str] = self.worker.transcribe(
             file_path,